Fix yad version check if statement
I was having stl report that my version of yad was '0'. Having checked
the code I found that the $YAD variable was just set to 'yad' which
meant that it would be checking for a file called yad in the current
directory. Swapping to command -v which will use the PATH variable to
determine if yad is present.