Make usage of SteamTinkerLaunch clearer in Readme
It was not clear enough that SteamTinkerLaunch should be used as a launch option only for native Linux games. In the past it was probably okay to use the launch option for Proton games but lately there were more users reporting issues. This updates the Readme to make it clearer that the launch option is really only intended for native Linux games, and that the compatibility tool is the intended way of using SteamTinkerLaunch with Proton games.

Also removes the incorrect note that the launch option is unavailable on Flatpak; I asked and it is actually available (flathub/com.valvesoftware.Steam.Utility.steamtinkerlaunch#40).