Update user-installed language files
When a user has language files installed manually they will not update automatically. We look for language files in the STL system config dir (such as the STL prefix on Steam Deck) or the directory the script is currently running from to update the language file.
This fixes missing and outdated language strings on Steam Deck.