Improvements to local install support
On each run we now check if we have a SYSTEMSTLCFGDIR, which by default (outside of Steam Deck and Flatpak) is selected as a folder somewhere in /usr. For local installs this means by default STL will not find the local lang folders or collection folders. This attempts to fix this by checking what the correct SYSTEMSTLCFGDIR should be on each run and updating the global.conf file. It is called on each run so that if the SteamTinkerLaunch installation changes (like if a user moves between a local and global install) their global config file will be updated.

This may not be perfect, but in testing it should be good enough for now. It should not break existing installs and should only help users with local non-root installs (i.e., ProtonUp-Qt users).