Add warning message when in Flatpak on Steam Deck
This is incorrect usage, but several users seem to have tried to do it.
No idea yet if this code will even work, I'm not even sure Flatpak STL can correctly run the probe commands that STL does for Steam Deck. But if it does work, it should run and show the error dialog post-install on Steam Deck.