minor
Fixes for shellcheck after the changes to the vortex list-supported improvements
Add small comments to please Bash parsers that incorrectly highlight parts of the script. Some
of the highlighters freak out with the more complex patterns, so added some comments to "escape"
some of the complex quote patterns that confuse it.