HMM: Add 1-click install support
Mods for each supported HMM game can now be downloaded and installed using the 1-click install button on GameBanana. There is also a command line for this with 'steamtinkerlaunch hmm u <url>'.
Also adds creation of a desktop file which will put SteamTinkerLaunch in the application menu for users, allowing them to install it much more easily.