HMM: Workaround Proton version not being found
This is a hack fix for now that falls back to Experimental (or stable) when GE-Proton cannot be found.
We need to fix the finding of custom GE releases if STL ever found GE-Proton7-latest at compatibilitytools.d, because otherwise
even if it's not there anymore, STL will still look for it for some reason because it still stores it in ProtonCSV.txt.

Probably will be an annoying fix eventually, but this should help users on SteamOS right now.