HMM: Don't quit installation on GitHub ping fail
The ping check was put into place as if we can't ping GitHub, it was assumed it was down and so we can't download.
However, issues have been reported where some domains cannot be reached even if they are still up.
For HMM specifically, this was reported in #677.

Instead of exiting the dlLatestHMM function when we fail to ping GitHub, we attempt to continue. Logging is in place to indicate that
downloading and installation may fail (e.g. if a user is offline, even steps like Winetricks may fail).
This is still experimental and may cause a myriad of other issues, or may be fine. We will see :-)