Ensure tray icon is always removed on close
Sometimes the tray icon gets left behind on exit, e.g. if STL crashes. When the program gets an EXIT signal, we always call a cleanup function that sends a kill signal to the tray icon PID, if the tray icon is enabled.
This should ensure the tray icon does not linger.