One-Time Run: Add option to use custom executable's directory as working directory
This allows some tools like the Committee of Zero STEINS;GATE Steam Patch to run, as it needs to be executed from the same directory as the installer.
Also adds an option to specify a custom working directory for a custom executable, in cases where the script's directory and the custom executable's directory are both not desirable.