Add option to force use Proton with a custom command
In some instances, the logic for checking a Window program can fail. This is used with custom commands to check if we should use Proton/Wine or run it as a native executable.
This adds a checkbox on the game menu to force use Proton with a custom command, which should fix instances where a Windows executable is incorrectly detected as a Linux executable and is not ran correctly.

For #710.