Update 'list' Command (#728)
* Command Line: Refactor list command

Previously the list command only listed out a bunch of AppIDs, now in addition to listing the AppIDs it can also
list the game name, game path, and all three.

The help page still needs updated to reflect these changes. There is also an issue with getting the name from
the App Manifest, which was implemented outside of this PR. xargs, which is used to strip trailing whitespace
from the game names fetched from the App Manifest, doesn't like when
game names have 'unmatched' single quotes in them, e.g. "Shantae and the
Pirate's Curse"

* Command Line: Fix path option for list command

* Command Line: Fix getGameDir not finding soundtracks

* Command line: Update list command usage

* Command Line: Add option to only show path for getgamedir command