Custom command: Pass GameScope, GameMode and MangoHUD when ONLY_CUSTOMCMD
When we enable ONLY_CUSTOMCMD, now GameScope, GameMode and MangoHUD will be enabled.
Still need to verify that we do not incorrectly do this on Steam Deck, just to be safe.
Also still does not apply to native games.
Finally, GameScope does not seem to pass its arguments yet, just the path to GameScope e.g. /usr/bin/gamescope