GameScope: Fix issue with some argument values not being parsed correctly
If there was a single-dash flag after a double-dash flag,
for example '--cursor /path/to/cursor.png -U --sharpness-value 5',
then the cursor path would contain the following single-dash flag(s)
up until the next double dash flag, meaning the path would be read as
invalid.

This fixes the issue. It was not introduced in this PR but has been
fixed!