Command Line: Add count to list function
Now by default the 'full' option will end byshowing a count of all owned/installed AppIDs as the 'Game Count'
This is only accurate to a point as it may count things like DLCs, but it is a nice touch nonetheless.

There is also a separate 'count' option added as part of this, where as well as ID, name, and path, 'count'
can be passed where only the count of the owned/installed games will be displayed.