Command line: Make getCompatData more robust
Instead of falling over if no symlink is found in the STL symlink folder,
use getGameDir to try and find the compatdata directory either in the
game library folder or the Steam root library folder.

Also fixes getCompatData falling over if a game has a semicolon in the
name, i.e. 'ROBOTICS;NOTES'