One-Time Run: Add support for Native Linux executables
This should include scripts, AppImages, native binaries, etc.

The logic used to detect native Linux executables is the same as for custom commands,
which means that it's not fool-proof. For custom commands, we have an option to
force Proton with custom executables, should the check be incorrect. Currently one
time run does not have support for this, but it is planned to be added before this
feature is merged.

The other remaining issue is that we still require a STEAM_COMPAT_DATA_PATH, as the
setOneTimeRunVars function still looks for this, and we rely on having this for
Proton games. Instead, we need to change the logic to return a warning instead of
an error, and then only check for this path if we're using a Wine executable.