One-Time Run: Refactor how arguments are passed
Instead of the crazy repeated conditionals, we now build an array of arguments to pass to commandlineOneTimeRun from OneTimeRunGui.
Proton cannot be forced yet, but this is the first step to allowing the --forceproton flag.

This should also be tested some more. I tested briefly and the arguments
seem to work correctly, but more testing is needed."