Vortex: Add option to use Steam Linux Runtime
Vortex runs with Proton, but doesn't run inside of the SLR. This may be
causing compatibility issues on Steam Deck.

This PR adds a new option to the Vortex options in the Global Menu to
enable the Steam Linux Runtime with Vortex. The changes required to
implement this were non-trivial, but a combination of Proton and native
games were tested before and after launching Vortex from the Global Menu
(to ensure SLRCMD from the Vortex launch was not being re-used for the
game launch). Everything in testing appears to still work fine.

The `setSLRReap` function was changed significantly to enable this
functionality. It can now take an option to force Proton (instead of
falling back to ISGAME=3 which causes native Linux SLR to be selected).
It can also take a custom Proton version to use, so we don't use
RUNPROTON by default. This means we can pass the Vortex version to
getSLRReap and fetch the correct Steam Linux Runtime specified in its
toolmanifest.

This implementation has the issue of very old compatibility tools which
pass the SLR to use from the launch command will probably not work, as
Vortex cannot detect this. However, recent Vortex versions require an
up-to-date Wine anyway, so I doubt this is a problem.

The changes made to setSLRReap here should make it more straightforward
to integrate with One-Time Run, MO2, and any other third-party
application. It could also help with a future custom command overhaul
where custom-commands could have a dedicated Proton version separate
from the game (#625).