Winetricks: Create local symlink for systemwide Winetricks
By creating a symlink to a local Winetricks, this may help the SLR use
systemwide Winetricks without us having to download Winetricks when
we want to use in the SLR.

This may help some cases particularly on SteamOS where some Winetricks
require the SLR to install properly i.e. dotnetdesktop6 for Vortex.