One-Time Run: Add option to reset entries back to default
Clears all One-Time Run related variables in the per-game config
to be back to their default values. Blank values essentially
reflect the defaults, as blank is false for checkboxes, none for
paths, and the Proton version will be handled appropriately too.

Also adds 'STEAM_COMPAT_CLIENT_INSTALL_PATH=', which the
Proton script now seems to require to launch with Proton.