Add Non-Steam Game: Fix Non-Steam AppID Generation
We need to generate a 4byte little endian hex from a signed 32bit integer.
This is what gets written to the shortcuts.vdf file. Steam will reject
the AppID if it is not in this format, and will overwrite it.

This fixes the AppID generation, allowing us to 'predict' the AppID
and will allow for implementing Non-Steam Game artwork selection.