SpecialK: Only return artifact hash for last successful workflow
Should fix SpecialK Nightly failing when the last workflow failed.
nightly.link only shows the artifacts for the lat successful
workflow, so the latest hash we got from a failed build would not
match the hash of the artifacts on nightly.link, so we coudln't
parse out the download URL.

Should resolve #975.