Fix generating Native Steam Linux Runtime command when it is not installed
Native games will crash if 'Steam Linux Runtime 1.0 (Scout)' is not installed,
this is because SteamTinkerLaunch incorrectly tries to fall back to generating
a wrong command and does not properly return the empty SLRCMD variable. This
leads to generating the launch command incorrectly, as STL thinks the SLR
is installed and is valid.

To fix this, we do an extra check to ensure the Proton SLR command is only
generated when we're using the Proton SLR, and which avoids us falling
back on the Proton SLR codepath when NATIVE_SLRCMD is not set. It is
incorrect to assume that because the NATIVE_SLRCMD is not set that we're
using Proton, we should only fall back to using PROTON_SLRCMD if we actually
set the PROTON_SLRCMD earlier.