Steam Linux Runtime: Re-enable parsing Steam Linux Runtime from Compatibility Tool start command
Primarily (but not exclusively) allows native games that
have a Steam Linux Runtime selected (either by the user,
or by Valve as sometimes they force SLR 1.0 or 3.0 for
games) to launch with the SLR selected.

Previously, the SLR would be ignored entirely if it
came from the compatibility tool, but this causes
incompatibilities with games that REQUIRE a specific
SLR version (ex: CS2 requires SLR 3.0 and is selected
by Valve, the user doesn't have a choice)

This change allows STL to read the SLR coming from
the compat tool launch, so if a game is launched
with a specified SLR version, we will use this
instead of ignoring or building the start command
ourselves.