Game Menu: Add option to disable Native Linux Steam Linux Runtime coming from Compatibility Tool
If a native game is launched with a Steam Linux Runtime Compatibility Tool
selected, either by the user or Valve Testing, the Steam Linux Runtime will
come in the game launch command. However a user may want to ignore this
for any reason, so this adds a checkbox to allow users to ignore the
incoming Steam Linux Runtime and let SteamTinkerLaunch find it.