CI: Further refine workflow
Should now only run once. Either on push, or on PR if the branch is not from this repo,
as changes pushed to branches on the repo will always be executed anwyay. We only want
PR workflows to trigger when the branch is from a fork, and we will use push events
for branches within the repo.

I hope this works.