fix createVdfEntry broken since last commit
Now inserts properly, so we should be able
to insert the new localconfig.vdf properly.

This also means adding a new non-steam
game compat tool should work with ansg,
but we need to test this fully to make
sure we didn't cause a regression wehn
touching createVdfEntry.

Should be safe though :-)