Add Non-Steam Game: Allow 'files' directory when checking Valve Proton
Previously Valve Proton versions only had a 'dist' directory.
This was one way to tell if a Proton version was a Valve version
or a community flavour.

As of Proton 9.0 (including Experimental, Beta, and Hotfix), Valve
Proton versions use the 'files' folder instead of 'dist', so we can
no longer assume that 'files' indicates a community Proton version.

This commit changes our check to accept 'files' and 'dist' folders.
We technically don't need these checks, but they are in place as
a sanity check to confirm our Proton folder is somewhat correct.
If the Proton version was missing a folder with this name it would
be invalid, so it's an extra safety check.

The main way we will differentiate a Valve Proton version now is
based on whether it has a `compatibilitytool.vdf` file in its
directory, as only community Proton flavours have this. Valve
Proton versions store this in an internal Steam file called
`appinfo.vdf`.