lang: Don't create global config in `loadLanguage`
Don't attempt to save and load the global config in `loadLanguage` unless it exists.
Prevents `loadLanguage`, which is called very early, from creating the global config
before `setSteamPaths` has a chance to define some paths that are needed to create
the global config.

Most notably, this fixes Luxtorpeda and Roberta variables from being incorrect when
created too early by `loadLanguage`, as this function creates the global config
before `setSteamPaths` has a chance to define `STEAMCOMPATOOLS` which is needed
in order to build a valid path for `LUXTORPEDACMD` and `ROBERTACMD`.